backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
django-filter>=24.3,<25.0
drf-spectacular>=0.27,<0.28
python-decouple>=3.8,<3.9
psycopg2-binary>=2.9,<2.10
Pillow>=10.4,<10.5
cloudinary>=1.36,<2.0
reportlab>=4.0,<5.0
gunicorn>=22.0,<23.0
